refactor!: remove old logic
