--- lua/telescope/_extensions/frecency2.lua
@@ -1,9 +1,0 @@
-local frecency = require "frecency"
-
-return require("telescope").register_extension {
-  setup = frecency.setup,
-  exports = {
-    frecency2 = frecency.start,
-    complete = frecency.complete,
-  },
-}
